+++ CHANGELOG.md
@@ -1,5 +1,14 @@
 # JSON-goLD ChangeLog
 
+## v0.3.0 - 2017-12-03
+
+### Changed
+
+- Implemented URDNA2015 RDF dataset normalisation algorithm
+- Added the _algorithm_ field JSON-LD options (accepts 'URDNA2015' or 'URGNA2012' (default))
+- Complete rewrite of the normalisation algorithm
+- Added RDF Dataset Normalisation test suite from [here](https://github.com/json-ld/normalization)
+
 ## v0.2.0 - 2017-12-02
 
 ### Changed